feat: implement GitHub stats display in sidebar and clean up layout components
