+++ docs/src/shells/DocsLayoutShell.tsx
@@ -3,6 +3,7 @@ import { DocsLayout } from 'fumadocs-ui/layouts/docs';
 import { baseOptions } from '@/lib/layout.shared';
 import { source } from '@/lib/source';
 import DocsSearch from '../components/DocsSearch';
+import { GithubSidebarFooter } from '../components/github-sidebar';
 
 export default function DocsLayoutShell({ children }: { children: ReactNode }) {
   return (
@@ -11,7 +12,9 @@ export default function DocsLayoutShell({ children }: { children: ReactNode }) {
       {...baseOptions()}
       sidebar={{
         tabs: false,
+        footer: <GithubSidebarFooter />,
       }}
+      themeSwitch={{ enabled: false }}
     >
       <DocsSearch />
       {children}
